requirements.txt

pytest
requests~=2.32.3
allure-pytest

allure-python-commons~=2.14.0